+++ changelog.md
@@ -14,6 +14,7 @@
 * [Changed] Lobulus boundary and Central vein segmentation parameters updated
 * [Added] Images from Lobulus boundary and Central vein segmentation
 * [Added] Processing parameters saved to yaml and json file
+* [Added] Collapsed parameter sub-tree with complex parameters
 
 ## 0.4.*